+++ tests/test_include.py
@@ -8,7 +8,7 @@ import marshmallow_jsonapi
 from marshmallow_jsonapi import fields
 
 from flask_jsonapi import api
-from flask_jsonapi import resource_repositories
+from flask_jsonapi import resource_repository_views
 
 JSONAPI_HEADERS = {'content-type': 'application/vnd.api+json', 'accept': 'application/vnd.api+json'}
 
@@ -48,7 +48,7 @@ class ParentDetailRepository:
         )
 
 
-class ParentResourceRepositoryViewSet(resource_repositories.ResourceRepositoryViewSet):
+class ParentResourceRepositoryViewSet(resource_repository_views.ResourceRepositoryViewSet):
     schema = ParentSchema
     repository = ParentDetailRepository()
